# Conflicts:
#	plc4c/generated-sources/s7/src/data_item.c
#	plc4c/generated-sources/s7/src/transport_size.c
#	plc4go/protocols/s7/readwrite/model/TransportSize.go
#	plc4j/drivers/s7/pom.xml
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/events/S7CyclicEvent.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/S7HGeneratedDriverBase.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/configuration/S7Configuration.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/connection/S7HDefaultNettyPlcConnection.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/optimizer/S7Optimizer.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/protocol/S7HGeneratedDriverBase.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/protocol/S7HMuxImpl.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/protocol/S7HPlcConnection.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/protocol/S7ProtocolLogic.java
#	plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/utils/StaticHelper.java

+++ plc4j/drivers/s7/src/main/java/org/apache/plc4x/java/s7/readwrite/optimizer/S7Optimizer.java
@@ -24,7 +24,6 @@ import org.apache.plc4x.java.api.model.PlcTag;
 import org.apache.plc4x.java.api.value.PlcValue;
 import org.apache.plc4x.java.s7.readwrite.*;
 import org.apache.plc4x.java.s7.readwrite.context.S7DriverContext;
-import org.apache.plc4x.java.s7.readwrite.tag.S7StringTag;
 import org.apache.plc4x.java.s7.readwrite.tag.S7Tag;
 import org.apache.plc4x.java.s7.readwrite.MemoryArea;
 import org.apache.plc4x.java.s7.readwrite.TransportSize;
@@ -65,13 +64,13 @@ public class S7Optimizer extends BaseOptimizer {
         LinkedHashMap<String, PlcTag> curTags = new LinkedHashMap<>();
 
         for (String tagName : readRequest.getTagNames()) {
-                    
-            if ((readRequest.getTag(tagName) instanceof S7SzlTag) || 
+
+            if ((readRequest.getTag(tagName) instanceof S7SzlTag) ||
                 (readRequest.getTag(tagName) instanceof S7ClkTag)) {
                 curTags.put(tagName, readRequest.getTag(tagName));
                 continue;
             }
-            
+
             S7Tag tag = (S7Tag) readRequest.getTag(tagName);
 
             int readRequestItemSize = S7_ADDRESS_ANY_SIZE;
@@ -129,13 +128,13 @@ public class S7Optimizer extends BaseOptimizer {
 
     @Override
     protected List<PlcWriteRequest> processWriteRequest(PlcWriteRequest writeRequest, DriverContext driverContext) {
-        
+
         for (String tagName : writeRequest.getTagNames()) {
             if (writeRequest.getTag(tagName) instanceof S7ClkTag) {
-                return Collections.singletonList( writeRequest);
+                return Collections.singletonList(writeRequest);
             }
-        } 
-        
+        }
+
         S7DriverContext s7DriverContext = (S7DriverContext) driverContext;
         List<PlcWriteRequest> processedRequests = new LinkedList<>();
 
@@ -146,14 +145,14 @@ public class S7Optimizer extends BaseOptimizer {
 
         // List of all items in the current request.
         LinkedHashMap<String, TagValueItem> curTags = new LinkedHashMap<>();
-        
+
         for (String tagName : writeRequest.getTagNames()) {
             S7Tag tag = (S7Tag) writeRequest.getTag(tagName);
             PlcValue value = writeRequest.getPlcValue(tagName);
 
             int writeRequestItemSize = S7_ADDRESS_ANY_SIZE + 4/* Size of Payload item header*/;
             if (tag.getDataType() == TransportSize.BOOL) {
-                writeRequestItemSize += Math.ceil((double) tag.getNumberOfElements() / 8);
+                writeRequestItemSize += (int) Math.ceil((double) tag.getNumberOfElements() / 8);
             } else {
                 int length = 1;
                 if(tag instanceof S7StringTag)
